Fix clang (with ndk r13) different function prototype compile error.

+++ source/Irrlicht/Android/CAndroidAssetReader.cpp
@@ -36,11 +36,11 @@ CAndroidAssetReader::~CAndroidAssetReader()
 		AAsset_close(Asset);
 }
 
-size_t CAndroidAssetReader::read(void* buffer, size_t sizeToRead)
+s32 CAndroidAssetReader::read(void* buffer, u32 sizeToRead)
 {
 	int readBytes = AAsset_read(Asset, buffer, sizeToRead);
 	if ( readBytes >= 0 )
-		return size_t(readBytes);
+		return s32(readBytes);
 	return 0;	// direct fd access is not possible (for example, if the asset is compressed). 
 }
